refactor(Sections): extract class
- 새로운 구간을 추가할 대상을 찾는 책임이 커짐에 따라 별도의 클래스로 분리
